Solved issues with validInputIDs
I have had to add NA values to fill in the empty space when the number
of inputs is larger than the number of rows in functionNodes. I also now
allow for duplicates to be added to the matrix but these are removed
when sampling from the valid choices. The NA values are also removed